20230125-avro_three_ways/python-fastavro/requirements.txt
fastavro
python-dateutil
